Finished pathing
So you can put in a start location and a finish location and it will figure out the path in between them.

There is one bug: if there are two paths to the same place, and they both pass between the same intermediatery planet, then it will choose the first path regardless of whether it is the shortest or not. This is because each planet visited is recorded as 'hit' and the second time it goes past that planet, it ignores the planet.